[native demo] Extended native demo to native-mobile
