+++ compose/native/demo/src/uikitMain/kotlin/androidx/compose/native/demo/main.uikit.kt
@@ -1,0 +1,40 @@
+// Use `xcodegen` first, then `open ./SkikoSample.xcodeproj` and then Run button in XCode.
+package androidx.compose.native.demo
+
+import androidx.compose.ui.window.Application
+import kotlinx.cinterop.*
+import platform.UIKit.*
+import platform.Foundation.*
+
+fun main() {
+    val args = emptyArray<String>()
+    memScoped {
+        val argc = args.size + 1
+        val argv = (arrayOf("skikoApp") + args).map { it.cstr.ptr }.toCValues()
+        autoreleasepool {
+            UIApplicationMain(argc, argv, null, NSStringFromClass(SkikoAppDelegate))
+        }
+    }
+}
+
+class SkikoAppDelegate : UIResponder, UIApplicationDelegateProtocol {
+    companion object : UIResponderMeta(), UIApplicationDelegateProtocolMeta
+
+    @ObjCObjectBase.OverrideInit
+    constructor() : super()
+
+    private var _window: UIWindow? = null
+    override fun window() = _window
+    override fun setWindow(window: UIWindow?) {
+        _window = window
+    }
+
+    override fun application(application: UIApplication, didFinishLaunchingWithOptions: Map<Any?, *>?): Boolean {
+        window = UIWindow(frame = UIScreen.mainScreen.bounds)
+        window!!.rootViewController = Application("Compose/Native sample") {
+            myContent()
+        }
+        window!!.makeKeyAndVisible()
+        return true
+    }
+}
